# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: Questions
    description: Quiz question bank (public read, admin write).
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /questions:
    get:
      tags: [Questions]
      operationId: listQuestions
      summary: List all quiz questions (public)
      responses:
        '200':
          description: Question list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
    post:
      tags: [Questions]
      operationId: createQuestion
      summary: Create a quiz question (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionInput'
      responses:
        '200':
          description: Question created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden — not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Questions]
      operationId: deleteAllQuestions
      summary: Delete every question (admin only)
      responses:
        '200':
          description: Deleted count.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedCount'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden — not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /questions/seed:
    post:
      tags: [Questions]
      operationId: seedQuestions
      summary: Bulk-insert questions (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionSeedBody'
      responses:
        '200':
          description: Number inserted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsertedCount'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden — not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /questions/{id}:
    patch:
      tags: [Questions]
      operationId: updateQuestion
      summary: Update a quiz question (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionInput'
      responses:
        '200':
          description: Question updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden — not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Questions]
      operationId: deleteQuestion
      summary: Delete a quiz question (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedOne'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden — not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user, isAdmin]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
        isAdmin:
          type: boolean
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SubjectCode:
      type: string
      enum: [M1, M2, M3, M4]
    Question:
      type: object
      required: [id, subject_code, q_en, q_hi, options_en, options_hi, a, exp_en, exp_hi]
      properties:
        id:
          type: string
        subject_code:
          $ref: '#/components/schemas/SubjectCode'
        q_en:
          type: string
        q_hi:
          type: string
        options_en:
          type: array
          items:
            type: string
          minItems: 4
          maxItems: 4
        options_hi:
          type: array
          items:
            type: string
          minItems: 4
          maxItems: 4
        a:
          type: integer
          minimum: 0
          maximum: 3
        exp_en:
          type: string
        exp_hi:
          type: string
    QuestionInput:
      type: object
      required: [subject_code, q_en, q_hi, options_en, options_hi, a, exp_en, exp_hi]
      properties:
        subject_code:
          $ref: '#/components/schemas/SubjectCode'
        q_en:
          type: string
          minLength: 1
        q_hi:
          type: string
          minLength: 1
        options_en:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 4
          maxItems: 4
        options_hi:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 4
          maxItems: 4
        a:
          type: integer
          minimum: 0
          maximum: 3
        exp_en:
          type: string
        exp_hi:
          type: string
    QuestionList:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    QuestionSeedBody:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionInput'
    InsertedCount:
      type: object
      required: [inserted]
      properties:
        inserted:
          type: integer
    DeletedCount:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: integer
    DeletedOne:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true