# lib/api-zod/src/generated/types/deletedOne.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export const DeletedOneValue = {
  success: true,
} as const;
export type DeletedOne = typeof DeletedOneValue;