# lib/api-zod/src/generated/types/questionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectCode } from "./subjectCode";

export interface QuestionInput {
  subject_code: SubjectCode;
  /** @minLength 1 */
  q_en: string;
  /** @minLength 1 */
  q_hi: string;
  /**
   * @minItems 4
   * @maxItems 4
   */
  options_en: string[];
  /**
   * @minItems 4
   * @maxItems 4
   */
  options_hi: string[];
  /**
   * @minimum 0
   * @maximum 3
   */
  a: number;
  exp_en: string;
  exp_hi: string;
}