# lib/api-zod/src/generated/types/questionList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from "./question";

export interface QuestionList {
  questions: Question[];
}