# lib/api-zod/src/generated/types/questionSeedBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionInput } from "./questionInput";

export interface QuestionSeedBody {
  questions: QuestionInput[];
}