# lib/api-zod/src/generated/types/subjectCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type SubjectCode = (typeof SubjectCode)[keyof typeof SubjectCode];

export const SubjectCode = {
  M1: "M1",
  M2: "M2",
  M3: "M3",
  M4: "M4",
} as const;